adjust naming to be more explicite
