+++ metastats-infrastructure/src/main/groovy/life/qbic/metastats/PrepareMetaData.groovy
@@ -2,25 +2,25 @@ package life.qbic.metastats
 
 import life.qbic.metastats.datamodel.MetaStatsExperiment
 import life.qbic.metastats.datamodel.MetaStatsSample
-import life.qbic.metastats.filter.FilterExperimentData
+import life.qbic.metastats.filter.FilterExperimentDataInput
 import life.qbic.metastats.request.ExperimentDataOutput
 
 class PrepareMetaData implements ExperimentDataOutput {
 
-    FilterExperimentData filterExperimentDataInput
+    FilterExperimentDataInput filterExperimentDataInput
 
     /**
      * Guides how data given from ExperimentDataOutput is transferred towards the use case FilterExperiment data
      * @param filterDataInput
      */
-    PrepareMetaData(FilterExperimentData filterDataInput) {
+    PrepareMetaData(FilterExperimentDataInput filterDataInput) {
         filterExperimentDataInput = filterDataInput
     }
 
 
     @Override
     def metaDataForProject(List<MetaStatsSample> sampleMetadata, List<MetaStatsExperiment> experimentMetadata) {
-        filterExperimentDataInput.filterProjectMetaData(sampleMetadata, experimentMetadata)
+        filterExperimentDataInput.getProjectMetaData(sampleMetadata, experimentMetadata)
     }
 
 }
